projects/ai-document-processor/api/pyproject.toml
[project]
name = "ai-document-processor-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "anthropic>=0.44",
    "pymupdf>=1.25",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17",
    "Pillow>=11.0",
    "python-docx>=1.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.18",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]
